# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LatestRun": {
        "description": "The derived latest-run read model carried on a project.\n\nArgs:\n    capability_run_id: Identity of the project's most recent run.\n    status: That run's current status.\n    started_at: When that run started executing.\n    ended_at: When that run reached a terminal status, or `None` while\n        still running or paused.",
        "properties": {
          "capability_run_id": {
            "format": "uuid",
            "title": "Capability Run Id",
            "type": "string"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "enum": [
              "running",
              "paused",
              "succeeded",
              "degraded",
              "failed",
              "aborted",
              "interrupted"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "capability_run_id",
          "status",
          "started_at"
        ],
        "title": "LatestRun",
        "type": "object"
      },
...
      "PageMeta": {
        "description": "Pagination metadata attached to a paginated list response.\n\nArgs:\n    page: 1-indexed page number returned.\n    page_size: Number of items requested per page (server-capped at\n        `PAGE_SIZE_MAX`).\n    total_items: Total number of items across all pages.",
        "properties": {
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total_items": {
            "minimum": 0.0,
            "title": "Total Items",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total_items"
        ],
        "title": "PageMeta",
        "type": "object"
      },
...
      "Page_ProjectOut_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProjectOut"
            },
            "title": "Data",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "title": "Page[ProjectOut]",
        "type": "object"
      },
...
      "PlanDraft": {
        "description": "Draft or approved orchestration plan, as surfaced to the client.\n\nMirrors the runtime `OrchestrationPlan` field-by-field. Every field\nexcept `steps`/`ready` may be `None`/absent while drafting.\n\nArgs:\n    title: Short user-visible title for the run.\n    question: Refined evidence question.\n    scoping_notes: User-expressed scoping notes.\n    screening_criteria: Visible inclusion/exclusion criteria for screening.\n    backend_scope: Search backend scope.\n    scope_constraints: Optional recency and publisher-geography constraints.\n    search_effort: Acquisition effort rung.\n    analysis_depth: Analysis component and budget rung.\n    components: Discretionary orchestration components only.\n    component_rationale: Visible intent-fit rationale keyed by\n        discretionary component.\n    grouping_facets: Optional grouping facets, valid only when `group`\n        runs.\n    extract_profiles: Optional finding profile short names, valid only\n        when `extract` runs.\n    steering_mode: Steering mode for the run.\n    assumptions: Visible assumptions and open guesses.\n    expected_artefact_shape: Deterministic forecast derived from components.\n    time_band: Deterministic wall-clock band derived from the two axes.\n    steps: The composed chain, presentation-labelled, in execution order.\n    ready: Whether the draft has validated fail-closed into an\n        executable plan.",
        "properties": {
          "analysis_depth": {
            "anyOf": [
              {
                "enum": [
                  "landscape",
                  "standard",
                  "deep"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Analysis Depth"
          },
          "assumptions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Assumptions"
          },
          "backend_scope": {
            "anyOf": [
              {
                "enum": [
                  "academic_only",
                  "grey_lit_only",
                  "both"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Backend Scope"
          },
          "component_rationale": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Component Rationale"
          },
          "components": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "characterise",
                    "screen_full",
                    "select",
                    "extract",
                    "group"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Components"
          },
          "expected_artefact_shape": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Expected Artefact Shape"
          },
          "extract_profiles": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "iof",
                    "icf"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Extract Profiles"
          },
          "grouping_facets": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "intervention",
                    "outcome",
                    "population",
                    "barrier_theme",
                    "enabler_theme",
                    "mechanism_theme"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Grouping Facets"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Question"
          },
          "ready": {
            "default": false,
            "title": "Ready",
            "type": "boolean"
          },
          "scope_constraints": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScopeConstraintsDraft"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "scoping_notes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Scoping Notes"
          },
          "screening_criteria": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Screening Criteria"
          },
          "search_effort": {
            "anyOf": [
              {
                "enum": [
                  "rapid",
                  "standard",
                  "deep"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Search Effort"
          },
          "steering_mode": {
            "anyOf": [
              {
                "enum": [
                  "frequent",
                  "moderate",
                  "minimal",
                  "unattended"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Steering Mode"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/PlanStep"
            },
            "title": "Steps",
            "type": "array"
          },
          "time_band": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Time Band"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Title"
          }
        },
        "title": "PlanDraft",
        "type": "object"
      },
      "PlanOut": {
        "description": "Response body for `GET /api/v1/projects/{id}/plan`.\n\nArgs:\n    plan: The current plan (draft or approved).\n    version: Plan row version.\n    status: Plan status (e.g. `draft`, `approved`).",
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/PlanDraft"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "plan",
          "version",
          "status"
        ],
        "title": "PlanOut",
        "type": "object"
      },
      "PlanStep": {
        "description": "One user-visible step in the composed chain, for progress display.\n\nArgs:\n    label: Short user-visible step label (presentation; may change).\n    blurb: Longer user-visible step description (presentation; may\n        change).\n    stage: Composed-chain stage name the step corresponds to.",
        "properties": {
          "blurb": {
            "title": "Blurb",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "label",
          "blurb",
          "stage"
        ],
        "title": "PlanStep",
        "type": "object"
      },
...
      "PlanningTurnCreate": {
        "additionalProperties": false,
        "description": "Inbound body for `POST /api/v1/projects/{id}/planning-turns`.\n\nArgs:\n    message: The user's chat message for this planner turn.\n    client_turn_id: Caller-minted UUID making double-submit idempotent \u2014\n        resubmitting the same `client_turn_id` returns the same turn\n        rather than re-running the planner.",
        "properties": {
          "client_turn_id": {
            "format": "uuid",
            "title": "Client Turn Id",
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message",
          "client_turn_id"
        ],
        "title": "PlanningTurnCreate",
        "type": "object"
      },
      "PlanningTurnOut": {
        "description": "Response body for one planner turn.\n\nArgs:\n    reply: The planner's conversational reply for this turn.\n    plan: The full current draft plan.\n    suggestions: The planner's suggested answers to its clarifying\n        question, rendered as tappable quick replies. Empty when none.",
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/PlanDraft"
          },
          "reply": {
            "title": "Reply",
            "type": "string"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "title": "Suggestions",
            "type": "array"
          }
        },
        "required": [
          "reply",
          "plan"
        ],
        "title": "PlanningTurnOut",
        "type": "object"
      },
      "ProjectCreate": {
        "additionalProperties": false,
        "description": "Inbound body for `POST /api/v1/projects`.\n\nArgs:\n    name: Project display name, 1-200 characters. Outer whitespace is\n        stripped before the length constraint is applied\n        (`str_strip_whitespace`).\n    question: Optional initial evidence question.",
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProjectCreate",
        "type": "object"
      },
      "ProjectOut": {
        "description": "A project resource.\n\nArgs:\n    project_id: The project's identity.\n    name: Current display name.\n    question: Current evidence question, or `None` if not yet set.\n    status: Lifecycle status (`active` unless archived; no hard delete).\n    created_at: When the project was created.\n    updated_at: When the project row was last written.\n    archived_at: When the project was archived, or `None` if active.\n    latest_run: The derived latest-run read model, or `None` before any\n        run has been created.",
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "latest_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LatestRun"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "status": {
            "enum": [
              "active",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "name",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectOut",
        "type": "object"
      },
...
      "ScopeConstraintsDraft": {
        "description": "Draft mirror of the runtime `ScopeConstraints`.\n\nArgs:\n    published_after: Optional lower publication-date bound (`YYYY-MM-DD`).\n    published_before: Optional upper publication-date bound (`YYYY-MM-DD`).\n    publisher_country: Optional Overton publisher-country filter.\n    author_affiliation_countries: Optional OpenAlex author-affiliation\n        country filter (2-letter alpha codes).\n    country_group: Optional named group applied to both search backends.",
        "properties": {
          "author_affiliation_countries": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Author Affiliation Countries"
          },
          "country_group": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CountryGroupDraft"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "published_after": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Published After"
          },
          "published_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Published Before"
          },
          "publisher_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Publisher Country"
          }
        },
        "title": "ScopeConstraintsDraft",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Policy Atlas API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/projects": {
      "get": {
        "description": "List the authenticated user's projects with derived latest-run state.",
        "operationId": "list_projects_api_v1_projects_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "active",
              "enum": [
                "active",
                "archived",
                "all"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProjectOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "description": "Create one active project owned by the authenticated subject.",
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/plan": {
      "get": {
        "description": "Return the durable approved plan or the surviving local draft session.",
        "operationId": "get_plan_api_v1_projects__project_id__plan_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Plan",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/v1/projects/{project_id}/planning-turns": {
      "post": {
        "description": "Advance one project's real planner conversation once per client turn id.",
        "operationId": "create_planning_turn_api_v1_projects__project_id__planning_turns_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanningTurnCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningTurnOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Planning Turn",
        "tags": [
          "planning"
        ]
      }
    },
...
  }
}